requirements.txt
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.5
click==8.0.1
idna==3.2
joblib==1.0.1
nltk==3.6.2
pysrt==1.1.2
regex==2021.8.28
requests==2.26.0
tqdm==4.62.2
urllib3==1.26.6
pyenchant==3.2.1
